Add subfolders to categorize some modules.
The main purpose behind this is to organize the arguably crowded root folder of the repository. The utilizes project relocation to allow projects to be specified in a sub directory.

Currently there are 4 categories added by this PR:
 - client
 - deprecated
 - experimental (may be omitted, only applies to fully experimental modules)
 - updating (may be omitted)

The `client` subdirectory contains modules which are only enabled on a Minecraft client.

The `deprecated` subdirectory is for deprecated modules.

The `experimental` subdirectory holds experimental modules.

Finally the `updating` subdirectory holds modules which need to be updated to the latest version.

The location of a module follows this order:

deprecated - If a module is deprecated, it will always be in deprecated subdirectory

experimental - If a module is experimental.

client - If a module is only available on a Minecraft client

root directory - If the module does not meet any of the above criteria.